+++ src/Api.php
@@ -137,6 +137,41 @@ class Api
 
 
     /**
+     * Send a group of photos or videos as an album.
+     *
+     *  *
+     * <code>
+     * $params = [
+     *   'chat_id'      => '',
+     *   'media'              => [],
+     *   'disable_notification'           => false,
+     *   'reply_to_message_id'          => 0,
+     * ];
+     * </code>
+     *
+     * @link https://core.telegram.org/bots/api#sendMediaGroup
+     *
+     * @param array    $params
+     *
+     * @var int|string $params ['chat_id']
+     * @var array     $params ['media']
+     * @var bool     $params ['disable_notification']
+     * @var Integer       $params ['reply_to_message_id']
+     *
+     * @return On success, an array of the sent Messages is returned.
+     */
+    public function sendMediaGroup(array $params){
+        $response = $this->post('sendMediaGroup', $params);
+
+        return $this->prepareResponse(function (TelegramResponse $response)
+        {
+            return collect($response->getResult())->map(function ($message) {
+                return new Message($message);
+            });
+        }, $response);
+    }
+
+    /**
      * Returns the TelegramClient service.
      *
      * @return TelegramClient
@@ -1486,7 +1521,7 @@ class Api
     }
 
     /**
-     * Pin a message in a supergroup.
+     * Pin a message in a supergroup or a channel.
      *
      * The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
      *
@@ -1510,13 +1545,13 @@ class Api
     }
 
     /**
-     * Unpin a message in a supergroup chat.
+     *  unpin a message in a supergroup or a channel.
      *
      * The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
      *
      * @link https://core.telegram.org/bots/api#unpinchatmessage
      *
-     * @param array $params
+     * @param array $params $params ['chat_id']
      *
      * @var int|string $params ['chat_id']
      *
@@ -2078,6 +2113,10 @@ class Api
         }
     }
 
+
+
+
+
     /**
      * Sends a request to Telegram Bot API and returns the result.
      *
@@ -2328,4 +2367,6 @@ class Api
             call_user_func_array($this->onRejected, [$response, $elapsedTime]);
         }
     }
+
+
 }