+++ changelog.md
@@ -1,3 +1,5 @@
+...
+
 # Version 0.5.0
 
 - Move some logic from `Api` to `TelegramRequest`, so that `TelegramRequest::params` represents the params passed to `Api` methods.